Introduce our first systemd service config file
Notifies us on keyboard connect, and automatically rebinds caps lock ->
escape. This is better than putting it in a .bash_profile or i3 config
because it runs in a periodic loop, meaning it catches cases where we
plug in the keyboard _after_ things have started.

It also includes a notification. In the current scheme of things, that
means it reminds us to use Fn+Q key combo to rebind Esc to tilde at the
keyboard level (not software, I think).