Fix issue where window names would be empty when using lsp popups
We were doing something dumb in the config when the titlestring option
did the trick. I think we probably had a reason for doing it so
complicatedly, but I do not recall why.

Also, the actual code we were using to set titlestring was overly
complicated. We must have been really tired when trying to set the
titlestring last time.